feat(codegen,ph-cli): move migrate templates to codegen templates dir
