+++ packages/codegen/src/templates/boilerplate/document-models.ts
