+++ packages/codegen/src/templates/boilerplate/editors.ts
