+++ packages/codegen/src/templates/boilerplate/index.html.ts
