+++ packages/codegen/src/templates/boilerplate/index.ts
