+++ packages/codegen/src/templates/boilerplate/package.json.ts
