+++ packages/codegen/src/templates/boilerplate/tsconfig.json.ts
