+++ packages/codegen/src/templates/index.ts
@@ -1,3 +1,9 @@
+export * from "./boilerplate/document-models.js";
+export * from "./boilerplate/editors.js";
+export * from "./boilerplate/index.html.js";
+export * from "./boilerplate/index.js";
+export * from "./boilerplate/package.json.js";
+export * from "./boilerplate/tsconfig.json.js";
 export * from "./document-editor/editor.js";
 export * from "./document-editor/module.js";
 export * from "./document-model/actions.js";
@@ -34,4 +40,3 @@ export * from "./drive-editor/components/FolderTree.js";
 export * from "./drive-editor/components/NavigationBreadcrumbs.js";
 export * from "./drive-editor/config.js";
 export * from "./drive-editor/editor.js";
-export * from "./unsafe-utils.js";
